BUGFIX: fixed implementations of "setter" methods that could lead to a crash.
Previously, the code was written as:

- (void)setName:(NSString *)name
{
	[_name release];
	_name = name;
	[_name retain];
}

If the value passed in, `name`, is actually the same string as `_name`, then this will lead to a crash or unexpected behavior in the `[_name retain]` line, because _name will already have been deallocated. Therefore, we first retain `name`, the value passed in, before releasing the old value.