CHMDocument: greatly simplified search interface and implementation.
Previously, there were 5 search related methods, and direct manipulation of the search-related menu items. I've simplified this by defining a `CHMDocumentSearchMode` enum which defines the current mode for the document, and consolidated the interface to 2 methods: `changeSearchMode:` and `search:`. Rather than manipulating menu items directly, turning them on or off, I added a `validateMenuItem:` method to do it only when needed.

Updated .gitattributes to allow git to diff the UTF-16 strings files properly. This requires the addition of the following lines to be added to your global `~/.gitconfig` file:

	[diff "localizablestrings"]
	textconv = "iconv -f utf-16 -t utf-8"

See http://stackoverflow.com/a/21020607/277952 for more info.