CHMDocument et. al.: greatly simplified and improved the text encoding/custom text encoding implementation.
Previously, there was kind of a convoluted (no offense ;-) way of dealing with a documents default encoding and with its custom encoding. The document merely stored a custom tag, and then to change the data using the custom encoding, would have to ask the menu controller for the IANA encoding name associated with the custom tag. Also, manipulation of the text encoding menu items was done directly, rather than indirectly, by using a `validateMenuItem:` call.

To fix this, the following changes:

CHMTextEncodingMenuController:
	This class has been eliminated and its functionality rolled over into the app controller.

CHMAppController:
	Replaces the role that CHMTextEncodingMenuController used to play in setting up the encoding menu at app launch.

	The menu items created have their tag set to the NSStringEncoding enum value associated with the IANA encoding name.
	The menu items created also have their `representedObject` set to the IANA string encoding name, which the document can use to get without having to call back to the app controller/menu controller to translate a tag to encoding name.

CHMDocument:
	Defines `encoding`, `customEncoding`, and `customEncodingName` properties which store their respective types.

	Since the readString() readTrimmedString() functions are called quite often, I changed them to take an NSStringEncoding directly instead of having to call nameToEncoding() each time.

	Updated the loadMetadata method to first get the encoding the document is supposed to be in before doing any of the following parsing of data.