Massive reworking and simplifying of `CHMDocument` class: it's now a higher-level class that works with a lower-level `CHMDocumentFile` model class.
IMO, the original `CHMDocument` tried to do too much, and confronted the problem at too low of a level. In other words, I don't think it should have to worry about what a `struct chmFile` is, or be reading individual bytes from the chm file. That ends up making for a massive class that does everything from as high a level as the updating the UI, to as low a level as reading individual bytes from a CHM file. Instead, the workload should be split up somewhat, by abstracting all of that lower-level stuff into a smarter model-level class that deals only with the CHM file. To that end, I've created a new class, `CHMDocumentFile`, which serves as the "smarter, lower-level model class".

CHMDocument:
	Now uses an instance of `CHMDocumentFile` as its model object.

	All of the search-related stuff (such as SearchKit SKIndexRefs, etc.) and the methods which create the search index, have been moved into `CHMDocumentFile`. The search interface is simplified again. I've done away with the `CHMScoredLinkItem` and `CHMSearchResults` classes in favor of a simpler approach that uses a `CHMSearchResult` class that wraps a `CHMLinkItem` instance.

	CHMDocument now sets itself to be the CHMDocumentFile's searchDelegate, and the search results returned are stored in an NSMutableArray.

	We now simplify things by storing the current `CHMLinkItem` we're viewing in a web view as `currentItem`. This makes setting the title of the tab and locating the current item in the outline view much easier (than having to reacquire it each time a page is loaded).

CHMDocumentFile:
	Added the ability to specify whether a document file should automatically prepare the search index upon creation. The default is YES, though it can be disabled if desired (for example, if the Quick Look generator switches to using this class instead of the chmox classes).

ITSSProtocol:
	Changed the protocol to use the `CHMDocumentFile` rather than the `CHMDocument` in the URLRequests.

CHMLinkItem:
	No longer create the `children` array in init, but wait until it's explicitly needed.