BUGFIX: reworked `CHMDocument` exporting-to-PDF feature; fixes crash.
Existing `CHMExporter` class and how `CHMDocument` made use of it was odd: the exporter was retaining itself, and releasing itself, and the crash was likely caused by memory management issues (zombie objects). I reworked the 2 classes to have each `CHMDocument` have its own `exporter` instance which it creates during the export process and destroys afterwords. The `CHMExporter` class now communicates to the document indirectly, by way of a `<CHMExporterDelegate>` protocol.

Updated some deprecated method calls to modern equivalents.

Improved the `CHMExporter` class somewhat in that it now writes to a temp PDF file that exists inside a temp folder that has a unique name. This should help avoid the obvious clobbering of the temp PDF file if two or more `CHMDocuments` were all to Export to PDF at the same time.

Removed the "Label" default text on the progress panel text field (in the CHMDocument.xib file).