Updated and improved the text encodings functionality.
Added additional possible text codings, especially those which correspond to the "expected" Windows-based encodings defined in CHMDocumentFile.

Changed the IANA encoding names from previous to the names that CFStringConvertEncodingToIANACharSetName() would return for all supported NSStringEncoding/CFStringEncodings. For example, instead of "LATIN1" and "CP1256", now we use "iso-8859-1" and "windows-1256".

Changed the name of the "Default" encoding menu item to "Automatic " and the name of the automatically used text encoding as defined in the CHM file. For example, "Automatic Western (Windows Latin 1)". The text encoding name is colored in gray to differentiate it from the word Automatic.