CHMKit: updated classes, ivars, and method names to clarify whether they work with `CHMArchiveItem` or `CHMLinkItem`.
Since I added the `CHMArchiveItem` class, which is similar in name to `CHMLinkItem`, it would be unclear which type of "Item" ivars like `items`, `currentItem`, or methods like `itemAtPath:` were referring to. To help clarify, I've made the method names and ivars more explicit. For example:

@property (nonatomic, retain) CHMLinkItem *currentItem;
becomes
@property (nonatomic, retain) CHMLinkItem *currentLinkItem;

The following methods:
- (NSData *)dataForObjectAtPath:(NSString *)aRelativePath relativeToItem:(CHMLinkItem *)anItem;
- (CHMLinkItem *)itemAtPath:(NSString *)aPath;

become:
- (NSData *)dataForObjectAtPath:(NSString *)aRelativePath relativeToLinkItem:(CHMLinkItem *)anItem;
- (CHMLinkItem *)linkItemAtPath:(NSString *)aPath;

Also updated the CHMSpotlightImporter and CHMQuickLook submodules to do the same, and to work with these changes.