CHMKit: BUGFIX: improved handling of CHMArchiveItem's chm_enumerate() call for malformed CHM files.
This patch changes the behavior for when, in CHMArchiveItem, chm_enumerate() returns 0 while enumerating the contents of the CHM file. I've encountered some CHM files that have more than one page entry, with the first page entry being valid, and the second page entry indicating a negative offset, which doesn't make sense. This would cause the enumerator to return an error, and I would abort the creation of the root CHMArchiveItem and return nil. Instead, I've changed the behavior to the console as a NOTE, but continue on with the creation of the root CHMArchiveItem.