Improved the Select Next Tab/Select Previous Tab menu functionality.
I changed the method that the Select Next Tab and Select Previous Tab menu items call from `selectNextTabViewItem:` and `selectPreviousTabViewItem:` to `chmDocumentSelectNextTabViewItem:` and `chmDocumentSelectPreviousTabViewItem:`. Prior to this, the tab view in the CHMDocument was becoming the first responder to the `select*TabViewItem:`  methods, and preventing the CHMDocument from receiving the `validateMenuItem:` call for them. By changing them to have a `chmDocument` prefix, the document becomes the first responder and can successfully disable the menu items when only one tab is present.

I also improved the functionality by allowing both menu commands to cycle through to the first or last tab (for example, when the last tab view item is selected, and you choose Select Next Tab, it will cycle back and select the first tab view item).