Fixed provisional load errors when opening bookmarks that reference CHMDocuments that aren't currently open in the app.
Prior to this patch, when you opened a bookmark that was referencing a document that wasn't currently open, the following would happen:
	1) Document controller would create and open the document, which would begin to load the home page (or last path if there was one saved)
	2) Bookmark controller would then tell the document to load the bookmarked URL before it had finished loading the initial page
	3) Document would log a provisional load failure error for the page load cancelation.

While this is not a huge deal, I revamped the way bookmarks are loaded. There's now a custom NSDocumentController subclass, CHMDocumentController, that offers a `openDocumentWithContentsOfURL:loadBookmark:error:` method to handle opening of documents which need to load bookmarks. The BookmarkController now checks to see whether there's a document open that corresponds to bookmark's target, and if it is, calls `loadURL:`, otherwise, it uses CHMDocumentController's `openDocumentWithContentsOfURL:loadBookmark:error:` method, passing it the bookmark to open. The CHMDocumentController and CHMDocument now have a `pendingBookmarkToOpen` property which is used to store the bookmark to be opened. In CHMDocument's `windowControllerDidLoadNib:` method, the document now checks to see whether their is a pending bookmark to be opened, and if so, loads that URL, otherwise loads the home page or last saved path, if available.