CHMDocument: improved sidebar view split view implementation.
I now use NSUserDefault's registerDefaults: to register the default values for the sidebar width and the text size multiplier.

Prior to this patch, the sidebar could be opened so wide as to completely collapse the main web view part of the document window, which doesn't make sense. Also, when opening a new document, the sidebar might be restored to a width that was greater than the width of the main web view, which also doesn't make much sense.

To fix this, I've implemented several of the <NSSplitViewDelegate> protocol methods to better constrain the sidebar view to have a maximum and minimum size.

I also fixed a possible bug that failed to take into account the width of the dividerThickness when setting the position of the divider when restoring the sidebar. This would result in the sidebar width being reduced by the splitView's dividerThickness (1) each time the sidebar was hidden and restored.